backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-genai
firebase-admin
chromadb
python-multipart
python-dotenv
